clients/mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.62.0",
    "langchain>=1.2.9",
    "langchain-google-genai>=4.2.0",
    "langchain-mcp-adapters>=0.2.1",
    "langgraph>=1.0.8",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.1",
]
